+++ release.config.js
@@ -3,8 +3,12 @@ const package = 'systemlink-grafana-plugins';
 module.exports = {
     branches: ['main'],
     plugins: [
-        '@semantic-release/commit-analyzer',
-        '@semantic-release/release-notes-generator',
+        ["@semantic-release/commit-analyzer", {
+            preset: "conventionalcommits"
+        }],
+        ["@semantic-release/release-notes-generator", {
+            preset: "conventionalcommits"
+        }],
         '@semantic-release/changelog',
         '@semantic-release/npm',
         ['@semantic-release/exec', {