+++ src/datasources/tag/TagDataSource.test.ts
@@ -1,8 +1,18 @@
-import { MockProxy } from 'jest-mock-extended';
-import { TagDataSource } from './TagDataSource';
-import { setupDataSource, createFetchError, getQueryBuilder, defaultQueryOptions } from 'test/fixtures';
 import { BackendSrv, TemplateSrv } from '@grafana/runtime';
-import { TagHistoryResponse, TagQuery, TagQueryType, TagsWithValues } from './types';
+import { DeepPartial } from 'core/types';
+import { MockProxy } from 'jest-mock-extended';
+import _ from 'lodash';
+import {
+  createFetchError,
+  createFetchResponse,
+  defaultQueryOptions,
+  getQueryBuilder,
+  mockTimers,
+  requestMatching,
+  setupDataSource,
+} from 'test/fixtures';
+import { TagDataSource } from './TagDataSource';
+import { TagQuery, TagQueryType, TagWithValue } from './types';
 
 let ds: TagDataSource, backendSrv: MockProxy<BackendSrv>, templateSrv: MockProxy<TemplateSrv>;
 
@@ -11,10 +21,13 @@ beforeEach(() => {
 });
 
 const buildQuery = getQueryBuilder<TagQuery>()({ type: TagQueryType.Current, workspace: '' });
+mockTimers();
 
 describe('testDatasource', () => {
   test('returns success', async () => {
-    backendSrv.get.calledWith('/nitag/v2/tags-count').mockResolvedValue(25);
+    backendSrv.fetch
+      .calledWith(requestMatching({ url: '/nitag/v2/tags-count' }))
+      .mockReturnValue(createFetchResponse(25));
 
     const result = await ds.testDatasource();
 
@@ -22,7 +35,9 @@ describe('testDatasource', () => {
   });
 
   test('bubbles up exception', async () => {
-    backendSrv.get.calledWith('/nitag/v2/tags-count').mockRejectedValue(createFetchError(400));
+    backendSrv.fetch
+      .calledWith(requestMatching({ url: '/nitag/v2/tags-count' }))
+      .mockReturnValue(createFetchError(400));
 
     await expect(ds.testDatasource()).rejects.toHaveProperty('status', 400);
   });
@@ -30,87 +45,124 @@ describe('testDatasource', () => {
 
 describe('queries', () => {
   test('tag current value', async () => {
-    backendSrv.post
-      .calledWith('/nitag/v2/query-tags-with-values', expect.objectContaining({ filter: 'path = "my.tag"' }))
-      .mockResolvedValue(createQueryTagsResponse('my.tag', '3.14'));
+    backendSrv.fetch
+      .calledWith(requestMatching({ url: '/nitag/v2/query-tags-with-values', data: { filter: 'path = "my.tag"' } }))
+      .mockReturnValue(createQueryTagsResponse());
 
     const result = await ds.query(buildQuery({ path: 'my.tag' }));
 
     expect(result.data).toEqual([
       {
-        fields: [{ name: 'value', values: ['3.14'] }],
-        name: 'my.tag',
+        fields: [{ name: 'my.tag', values: [3.14] }],
         refId: 'A',
       },
     ]);
   });
 
   test('applies query defaults when missing fields', async () => {
-    backendSrv.post.mockResolvedValue(createQueryTagsResponse('my.tag', '3.14'));
+    backendSrv.fetch.mockReturnValue(createQueryTagsResponse());
 
-    const result = await ds.query({ ...defaultQueryOptions, targets: [{ path: 'my.tag'} as TagQuery]});
+    const result = await ds.query({ ...defaultQueryOptions, targets: [{ path: 'my.tag' } as TagQuery] });
 
-    expect(result.data[0]).toHaveProperty('fields', [{ name: 'value', values: ['3.14'] }]);
+    expect(result.data[0]).toHaveProperty('fields', [{ name: 'my.tag', values: [3.14] }]);
   });
 
   test('uses displayName property', async () => {
-    backendSrv.post.mockResolvedValue(createQueryTagsResponse('my.tag', '3.14', 'My cool tag'));
+    backendSrv.fetch.mockReturnValue(createQueryTagsResponse({ properties: { displayName: 'My cool tag' } }));
 
     const result = await ds.query(buildQuery({ path: 'my.tag' }));
 
-    expect(result.data[0]).toEqual(expect.objectContaining({ name: 'My cool tag' }));
+    expect(result.data[0].fields[0]).toHaveProperty('name', 'My cool tag');
+  });
+
+  test('handles null tag properties', async () => {
+    backendSrv.fetch.mockReturnValue(createQueryTagsResponse({ properties: null }));
+
+    const result = await ds.query(buildQuery({ path: 'my.tag' }));
+
+    expect(result.data[0]).toHaveProperty('fields', [{ name: 'my.tag', values: [3.14] }]);
+  });
+
+  test('handles tag with no current value', async () => {
+    backendSrv.fetch.mockReturnValue(createQueryTagsResponse({}, null));
+
+    const result = await ds.query(buildQuery({ path: 'my.tag' }));
+
+    expect(result.data[0]).toHaveProperty('fields', [{ name: 'my.tag', values: [] }]);
   });
 
   test('multiple targets - skips invalid queries', async () => {
-    backendSrv.post
-      .mockResolvedValueOnce(createQueryTagsResponse('my.tag1', '3.14'))
-      .mockResolvedValueOnce(createQueryTagsResponse('my.tag2', 'foo'));
+    backendSrv.fetch
+      .calledWith(requestMatching({ data: { filter: 'path = "my.tag1"' } }))
+      .mockReturnValue(createQueryTagsResponse({ path: 'my.tag1' }));
+    backendSrv.fetch
+      .calledWith(requestMatching({ data: { filter: 'path = "my.tag2"' } }))
+      .mockReturnValue(createQueryTagsResponse({ path: 'my.tag2' }, { value: { value: '41.3' } }));
 
     const result = await ds.query(buildQuery({ path: 'my.tag1' }, { path: '' }, { path: 'my.tag2' }));
 
-    expect(backendSrv.post.mock.calls[0][1]).toHaveProperty('filter', 'path = "my.tag1"');
-    expect(backendSrv.post.mock.calls[1][1]).toHaveProperty('filter', 'path = "my.tag2"');
     expect(result.data).toEqual([
       {
-        fields: [{ name: 'value', values: ['3.14'] }],
-        name: 'my.tag1',
+        fields: [{ name: 'my.tag1', values: [3.14] }],
         refId: 'A',
       },
       {
-        fields: [{ name: 'value', values: ['foo'] }],
-        name: 'my.tag2',
+        fields: [{ name: 'my.tag2', values: [41.3] }],
         refId: 'C',
       },
     ]);
   });
 
-  test('throw when no tags matched', async () => {
-    backendSrv.post.mockResolvedValue({ tagsWithValues: [] });
+  test('current value for all data types', async () => {
+    backendSrv.fetch
+      .mockReturnValueOnce(createQueryTagsResponse({ datatype: 'INT', path: 'tag1' }, { value: { value: '3' } }))
+      .mockReturnValueOnce(createQueryTagsResponse({ datatype: 'DOUBLE', path: 'tag2' }, { value: { value: '3.3' } }))
+      .mockReturnValueOnce(createQueryTagsResponse({ datatype: 'STRING', path: 'tag3' }, { value: { value: 'foo' } }))
+      .mockReturnValueOnce(createQueryTagsResponse({ datatype: 'BOOLEAN', path: 'tag4' }, { value: { value: 'True' } }))
+      .mockReturnValueOnce(
+        createQueryTagsResponse({ datatype: 'U_INT64', path: 'tag5' }, { value: { value: '2147483648' } })
+      );
 
-    await expect(ds.query(buildQuery({ path: 'my.tag' }))).rejects.toThrow(
-      'my.tag'
+    const result = await ds.query(
+      buildQuery({ path: 'tag1' }, { path: 'tag2' }, { path: 'tag3' }, { path: 'tag4' }, { path: 'tag5' })
     );
+
+    expect(result.data.map(frames => frames.fields[0])).toEqual([
+      { name: 'tag1', values: [3] },
+      { name: 'tag2', values: [3.3] },
+      { name: 'tag3', values: ['foo'] },
+      { name: 'tag4', values: ['True'] },
+      { name: 'tag5', values: [2147483648] },
+    ]);
+  });
+
+  test('throw when no tags matched', async () => {
+    backendSrv.fetch.mockReturnValue(createFetchResponse({ tagsWithValues: [] }));
+
+    await expect(ds.query(buildQuery({ path: 'my.tag' }))).rejects.toThrow("No tags matched the path 'my.tag'");
   });
 
   test('numeric tag history', async () => {
     const queryRequest = buildQuery({ type: TagQueryType.History, path: 'my.tag' });
 
-    backendSrv.post
-      .calledWith('/nitag/v2/query-tags-with-values', expect.objectContaining({ filter: 'path = "my.tag"' }))
-      .mockResolvedValue(createQueryTagsResponse('my.tag', '3.14'));
+    backendSrv.fetch
+      .calledWith(requestMatching({ url: '/nitag/v2/query-tags-with-values', data: { filter: 'path = "my.tag"' } }))
+      .mockReturnValue(createQueryTagsResponse());
 
-    backendSrv.post
+    backendSrv.fetch
       .calledWith(
-        '/nitaghistorian/v2/tags/query-decimated-history',
-        expect.objectContaining({
-          paths: ['my.tag'],
-          workspace: '1',
-          startTime: queryRequest.range.from.toISOString(),
-          endTime: queryRequest.range.to.toISOString(),
-          decimation: 300,
+        requestMatching({
+          url: '/nitaghistorian/v2/tags/query-decimated-history',
+          data: {
+            paths: ['my.tag'],
+            workspace: '1',
+            startTime: queryRequest.range.from.toISOString(),
+            endTime: queryRequest.range.to.toISOString(),
+            decimation: 300,
+          },
         })
       )
-      .mockResolvedValue(
+      .mockReturnValue(
         createTagHistoryResponse('my.tag', 'DOUBLE', [
           { timestamp: '2023-01-01T00:00:00Z', value: '1' },
           { timestamp: '2023-01-01T00:01:00Z', value: '2' },
@@ -123,17 +175,16 @@ describe('queries', () => {
       {
         fields: [
           { name: 'time', values: [1672531200000, 1672531260000] },
-          { name: 'value', values: [1, 2] },
+          { name: 'my.tag', values: [1, 2] },
         ],
-        name: 'my.tag',
         refId: 'A',
       },
     ]);
   });
 
   test('string tag history', async () => {
-    backendSrv.post.mockResolvedValueOnce(createQueryTagsResponse('my.tag', 'foo'));
-    backendSrv.post.mockResolvedValueOnce(
+    backendSrv.fetch.mockReturnValueOnce(createQueryTagsResponse());
+    backendSrv.fetch.mockReturnValueOnce(
       createTagHistoryResponse('my.tag', 'STRING', [
         { timestamp: '2023-01-01T00:00:00Z', value: '3.14' },
         { timestamp: '2023-01-01T00:01:00Z', value: 'foo' },
@@ -146,9 +197,8 @@ describe('queries', () => {
       {
         fields: [
           { name: 'time', values: [1672531200000, 1672531260000] },
-          { name: 'value', values: ['3.14', 'foo'] },
+          { name: 'my.tag', values: ['3.14', 'foo'] },
         ],
-        name: 'my.tag',
         refId: 'A',
       },
     ]);
@@ -158,9 +208,9 @@ describe('queries', () => {
     const queryRequest = buildQuery({ type: TagQueryType.History, path: 'my.tag' });
     queryRequest.maxDataPoints = 1500;
 
-    backendSrv.post.mockResolvedValueOnce(createQueryTagsResponse('my.tag', '3'));
+    backendSrv.fetch.mockReturnValueOnce(createQueryTagsResponse());
 
-    backendSrv.post.mockResolvedValueOnce(
+    backendSrv.fetch.mockReturnValueOnce(
       createTagHistoryResponse('my.tag', 'INT', [
         { timestamp: '2023-01-01T00:00:00Z', value: '1' },
         { timestamp: '2023-01-01T00:01:00Z', value: '2' },
@@ -169,46 +219,70 @@ describe('queries', () => {
 
     await ds.query(queryRequest);
 
-    expect(backendSrv.post.mock.lastCall?.[1]).toHaveProperty('decimation', 1000);
+    expect(backendSrv.fetch.mock.lastCall?.[0].data).toHaveProperty('decimation', 1000);
   });
 
   test('replaces tag path with variable', async () => {
     templateSrv.replace.calledWith('$my_variable').mockReturnValue('my.tag');
-    backendSrv.post
-      .calledWith('/nitag/v2/query-tags-with-values', expect.objectContaining({ filter: 'path = "my.tag"' }))
-      .mockResolvedValue(createQueryTagsResponse('my.tag', '3.14'));
+    backendSrv.fetch
+      .calledWith(requestMatching({ url: '/nitag/v2/query-tags-with-values', data: { filter: 'path = "my.tag"' } }))
+      .mockReturnValue(createQueryTagsResponse());
 
     const result = await ds.query(buildQuery({ type: TagQueryType.Current, path: '$my_variable' }));
 
-    expect(result.data[0]).toHaveProperty('name', 'my.tag');
+    expect(result.data[0].fields[0]).toHaveProperty('name', 'my.tag');
   });
 
   test('filters by workspace if provided', async () => {
-    backendSrv.post.mockResolvedValueOnce(createQueryTagsResponse('my.tag', '3.14', undefined, '2'));
-    backendSrv.post.mockResolvedValueOnce(createTagHistoryResponse('my.tag', 'DOUBLE', []));
+    backendSrv.fetch.mockReturnValueOnce(createQueryTagsResponse({ workspace_id: '2' }));
+    backendSrv.fetch.mockReturnValueOnce(createTagHistoryResponse('my.tag', 'DOUBLE', []));
 
     await ds.query(buildQuery({ type: TagQueryType.History, path: 'my.tag', workspace: '2' }));
 
-    expect(backendSrv.post.mock.calls[0][1]).toHaveProperty('filter', 'path = "my.tag" && workspace = "2"');
-    expect(backendSrv.post.mock.calls[1][1]).toHaveProperty('workspace', '2');
+    expect(backendSrv.fetch.mock.calls[0][0].data).toHaveProperty('filter', 'path = "my.tag" && workspace = "2"');
+    expect(backendSrv.fetch.mock.calls[1][0].data).toHaveProperty('workspace', '2');
+  });
+
+  test('retries failed request with 429 status', async () => {
+    backendSrv.fetch.mockReturnValueOnce(createQueryTagsResponse());
+    backendSrv.fetch.mockReturnValueOnce(createFetchError(429));
+    backendSrv.fetch.mockReturnValueOnce(createTagHistoryResponse('my.tag', 'INT', []));
+
+    await ds.query(buildQuery({ type: TagQueryType.History, path: 'my.tag' }));
+
+    expect(backendSrv.fetch).toHaveBeenCalledTimes(3);
+  });
+
+  test('retries a maximum of three times', async () => {
+    backendSrv.fetch.mockReturnValueOnce(createQueryTagsResponse());
+    backendSrv.fetch.mockReturnValue(createFetchError(429));
+
+    await expect(ds.query(buildQuery({ type: TagQueryType.History, path: 'my.tag' }))).rejects.toHaveProperty(
+      'status',
+      429
+    );
+
+    expect(backendSrv.fetch).toHaveBeenCalledTimes(5);
   });
 });
 
-function createQueryTagsResponse(path: string, value: string, displayName?: string, workspace_id = '1'): TagsWithValues {
-  return {
+function createQueryTagsResponse(
+  tag?: DeepPartial<TagWithValue['tag']>,
+  current?: DeepPartial<TagWithValue['current']>
+) {
+  return createFetchResponse({
     tagsWithValues: [
-      {
-        current: { value: { value } },
-        tag: { path, properties: { displayName }, workspace_id },
-      },
+      _.defaultsDeep(
+        { tag, current },
+        {
+          current: { value: { value: '3.14' } },
+          tag: { datatype: 'DOUBLE', path: 'my.tag', properties: {}, workspace_id: '1' },
+        }
+      ),
     ],
-  };
+  });
 }
 
-function createTagHistoryResponse(
-  path: string,
-  type: string,
-  values: Array<{ timestamp: string; value: string }>
-): TagHistoryResponse {
-  return { results: { [path]: { type, values } } };
+function createTagHistoryResponse(path: string, type: string, values: Array<{ timestamp: string; value: string }>) {
+  return createFetchResponse({ results: { [path]: { type, values } } });
 }