+++ src/datasources/tag/components/TagQueryEditor.test.tsx
@@ -1,48 +1,41 @@
-import { screen } from '@testing-library/react';
+import { screen, waitForElementToBeRemoved } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { TagQueryEditor } from './TagQueryEditor';
+import { select } from 'react-select-event';
 import { setupRenderer } from 'test/fixtures';
-import { TagQuery, TagQueryType } from '../types';
 import { TagDataSource } from '../TagDataSource';
+import { TagQuery, TagQueryType } from '../types';
+import { TagQueryEditor } from './TagQueryEditor';
 
 const render = setupRenderer(TagQueryEditor, TagDataSource);
+const workspacesLoaded = () => waitForElementToBeRemoved(screen.getByTestId('Spinner'));
 
 it('renders with query defaults', async () => {
   render({} as TagQuery);
+  await workspacesLoaded();
 
   expect(screen.getByRole('radio', { name: 'Current' })).toBeChecked();
   expect(screen.getByLabelText('Tag path')).not.toHaveValue();
-  expect(screen.getByLabelText('Workspace')).not.toHaveValue();
+  expect(screen.getByRole('combobox')).toHaveAccessibleDescription('Any workspace');
 });
 
-it('renders with saved query values', async () => {
-  render({ type: TagQueryType.History, path: 'my.tag', workspace: '1' });
+it('renders with initial query and updates when user makes changes', async () => {
+  const [onChange] = render({ type: TagQueryType.History, path: 'my.tag', workspace: '1' });
+  await workspacesLoaded();
 
+  // Renders saved query
   expect(screen.getByRole('radio', { name: 'History' })).toBeChecked();
   expect(screen.getByLabelText('Tag path')).toHaveValue('my.tag');
-  expect(screen.getByLabelText('Workspace')).toHaveValue('1');
-});
+  expect(screen.getByText('Default workspace')).toBeInTheDocument();
 
-it('updates query when user types new path', async () => {
-  const [onChange] = render({ type: TagQueryType.Current, path: '', workspace: '' });
+  // Users changes query type
+  await userEvent.click(screen.getByRole('radio', { name: 'Current' }));
+  expect(onChange).toBeCalledWith(expect.objectContaining({ type: TagQueryType.Current }));
 
-  await userEvent.type(screen.getByLabelText('Tag path'), 'my.tag{enter}');
+  // User types in new tag path
+  await userEvent.type(screen.getByLabelText('Tag path'), '.test{enter}');
+  expect(onChange).toBeCalledWith(expect.objectContaining({ path: 'my.tag.test' }));
 
-  expect(onChange).toBeCalledWith(expect.objectContaining({ path: 'my.tag' }));
-});
-
-it('updates query when user selects new type', async () => {
-  const [onChange] = render({ type: TagQueryType.Current, path: '', workspace: '' });
-
-  await userEvent.click(screen.getByRole('radio', { name: 'History' }));
-
-  expect(onChange).toBeCalledWith(expect.objectContaining({ type: TagQueryType.History }));
-});
-
-it('updates query when user types new workspace', async () => {
-  const [onChange] = render({ type: TagQueryType.Current, path: '', workspace: '' });
-
-  await userEvent.type(screen.getByLabelText('Workspace'), '1234{enter}');
-
-  expect(onChange).toBeCalledWith(expect.objectContaining({ workspace: '1234' }));
+  // User selects different workspace
+  await select(screen.getByRole('combobox'), 'Other workspace', { container: document.body });
+  expect(onChange).toBeCalledWith(expect.objectContaining({ workspace: '2' }));
 });