+++ src/datasources/tag/components/TagQueryEditor.tsx
@@ -1,15 +1,17 @@
+import { QueryEditorProps, SelectableValue } from '@grafana/data';
+import { AutoSizeInput, RadioButtonGroup, Select } from '@grafana/ui';
+import { InlineField } from 'core/components/InlineField';
+import { enumToOptions, useWorkspaceOptions } from 'core/utils';
 import React, { FormEvent } from 'react';
-import { AutoSizeInput, RadioButtonGroup } from '@grafana/ui';
-import { QueryEditorProps } from '@grafana/data';
 import { TagDataSource } from '../TagDataSource';
 import { TagQuery, TagQueryType } from '../types';
-import { InlineField } from 'core/components/InlineField';
-import { enumToOptions } from 'core/utils';
 
 type Props = QueryEditorProps<TagDataSource, TagQuery>;
 
 export function TagQueryEditor({ query, onChange, onRunQuery, datasource }: Props) {
   query = datasource.prepareQuery(query);
+
+  const workspaces = useWorkspaceOptions(datasource);
 
   const onTypeChange = (value: TagQueryType) => {
     onChange({ ...query, type: value });
@@ -21,8 +23,8 @@ export function TagQueryEditor({ query, onChange, onRunQuery, datasource }: Prop
     onRunQuery();
   };
 
-  const onWorkspaceChange = (event: FormEvent<HTMLInputElement>) => {
-    onChange({ ...query, workspace: event.currentTarget.value });
+  const onWorkspaceChange = (option?: SelectableValue<string>) => {
+    onChange({ ...query, workspace: option?.value ?? '' });
     onRunQuery();
   };
 
@@ -32,14 +34,16 @@ export function TagQueryEditor({ query, onChange, onRunQuery, datasource }: Prop
         <RadioButtonGroup options={enumToOptions(TagQueryType)} value={query.type} onChange={onTypeChange} />
       </InlineField>
       <InlineField label="Tag path" labelWidth={14} tooltip={tooltips.tagPath}>
-        <AutoSizeInput minWidth={20} defaultValue={query.path} onCommitChange={onPathChange} />
+        <AutoSizeInput minWidth={20} maxWidth={80} defaultValue={query.path} onCommitChange={onPathChange} />
       </InlineField>
       <InlineField label="Workspace" labelWidth={14} tooltip={tooltips.workspace}>
-        <AutoSizeInput
-          minWidth={20}
+        <Select
+          isClearable
+          isLoading={workspaces.loading}
+          onChange={onWorkspaceChange}
+          options={workspaces.value}
           placeholder="Any workspace"
-          defaultValue={query.workspace}
-          onCommitChange={onWorkspaceChange}
+          value={query.workspace}
         />
       </InlineField>
     </>
@@ -53,6 +57,6 @@ const tooltips = {
 
   tagPath: `The full path of the tag to visualize. You can enter a variable into this field.`,
 
-  workspace: `The ID of the workspace to search for the given tag path. If left blank, the plugin
+  workspace: `The workspace to search for the given tag path. If left blank, the plugin
               finds the most recently updated tag in any workspace.`,
 };