+++ src/datasources/tag/types.ts
@@ -12,10 +12,11 @@ export interface TagQuery extends DataQuery {
 }
 
 export interface TagWithValue {
-  current: { value: { value: string } };
+  current: { value: { value: string } } | null;
   tag: {
+    datatype: string;
     path: string;
-    properties: { displayName?: string };
+    properties: { displayName?: string } | null;
     workspace_id: string;
   };
 }